Prototype game timer
Made a single scene GameTimer to get a countdown for the game to put pressure on the player to act and not sit idle. In it's current form, it will likely need changes. It's not hooked up to any scene change. I was just work shopping this one scene and it can be deleted safely without effecting the game.